requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
